modified:   workspaces/DummyRosterOne/DummyRoster.Common.EntityModel/Models/Form.cs

+++ workspaces/DummyRosterOne/DummyRoster.Common.EntityModel/Models/Form.cs
@@ -48,7 +48,7 @@ public partial class Form
     [Column(TypeName = "DATETIME")]
     public DateTime? PromisedDate { get; set; }
 
-    public double? ShippingCost { get; set; }
+    public decimal? ShippingCost { get; set; }
 
     [ForeignKey("CarrierId")]
     [InverseProperty("Forms")]
